+++ src/main/java/hu/jakab/ekkeencoprosampbackend/dto/measurementUnit/MeasurementUnitRequestDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MeasurementUnitRequestDTO {
     private Long id; // Nullable for creation, required for updates
@@ -19,7 +21,7 @@ public class MeasurementUnitRequestDTO {
 
     private Long baseUnitId; // ID reference to base unit (nullable)
 
-    private Double conversionFactor; // Factor to convert to base unit
+    private BigDecimal conversionFactor; // Factor to convert to base unit
 
     private String standardBody; // e.g., "SI", "ISO", "ASTM", "EPA"
 }
