+++ src/main/java/hu/jakab/ekkeencoprosampbackend/dto/measurementUnit/MeasurementUnitResponseDTO.java
@@ -1,6 +1,8 @@
 package hu.jakab.ekkeencoprosampbackend.dto.measurementUnit;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,7 +11,7 @@ public class MeasurementUnitResponseDTO {
     private String unitCode;
     private String description;
     private String unitCategory;
-    private Double conversionFactor;
+    private BigDecimal conversionFactor;
     private String standardBody;
     private MeasurementUnitBaseUnitDTO baseUnit;
     private LocalDateTime createdAt;
